move conversion logic to adjustNetworkTopologySpec in session.go.
