+++ pkg/scheduler/actions/allocate/allocate.go
@@ -177,14 +177,6 @@ func (alloc *Action) buildAllocateContext() *allocateContext {
 			continue
 		}
 
-		// Convert soft topology to hard topology with ClusterTopHyperNode tier as maxTier,
-		// so that soft-mode jobs reuse the hard-mode scheduling path without any HyperNode filtering.
-		// This is done here (once per job) rather than in allocateResources loop to avoid redundant
-		// conversion when jobs are pushed back into the queue.
-		if clusterTopHyperNode, exists := ssn.HyperNodes[framework.ClusterTopHyperNode]; exists {
-			convertSoftToHardTopology(job, clusterTopHyperNode.Tier())
-		}
-
 		worksheet := alloc.organizeJobWorksheet(job)
 		if worksheet.Empty() {
 			continue
@@ -675,58 +667,6 @@ func (alloc *Action) allocateResourcesForTasks(subJob *api.SubJobInfo, tasks *ut
 
 	stmt.Discard()
 	return nil
-}
-
-// convertSoftToHardTopology converts all soft network topology constraints in the job to hard mode.
-// Conversion strategy:
-//   - Job-level soft: converted with maxTier (ClusterTopHyperNode tier), achieving no HyperNode
-//     filtering (full soft affinity across the cluster).
-//   - SubGroupPolicy/SubJob-level soft: converted with the effective job-level tier (subJobMaxTier).
-//     If the job has a hard tier limit (either user-specified or from the job-level conversion above),
-//     subgroup soft affinity is bounded by that limit. This properly handles the mixed-mode scenario
-//     where job is hard but subgroup is soft: the subgroup prefers lower tiers but never exceeds
-//     the job's hard tier constraint.
-func convertSoftToHardTopology(job *api.JobInfo, maxTier int) {
-	if job.PodGroup == nil {
-		return
-	}
-
-	// Convert job-level soft topology to hard mode with maxTier.
-	if job.PodGroup.Spec.NetworkTopology != nil &&
-		job.PodGroup.Spec.NetworkTopology.Mode == scheduling.SoftNetworkTopologyMode {
-		klog.V(3).InfoS("Converting job-level soft topology to hard mode",
-			"job", job.UID, "maxTier", maxTier)
-		job.PodGroup.Spec.NetworkTopology.Mode = scheduling.HardNetworkTopologyMode
-		job.PodGroup.Spec.NetworkTopology.HighestTierAllowed = &maxTier
-		job.PodGroup.Spec.NetworkTopology.HighestTierName = ""
-	}
-
-	// Determine the effective maxTier for SubGroupPolicy/SubJob conversion.
-	// If the job has an effective tier limit (from hard mode or job-level soft→hard conversion above),
-	// subgroup soft affinity must be bounded by it. Otherwise, fall back to the cluster-wide maxTier.
-	subJobMaxTier := maxTier
-	if job.PodGroup.Spec.NetworkTopology != nil &&
-		job.PodGroup.Spec.NetworkTopology.HighestTierAllowed != nil {
-		subJobMaxTier = *job.PodGroup.Spec.NetworkTopology.HighestTierAllowed
-	}
-
-	// Convert SubGroupPolicy-level soft topology with subJobMaxTier.
-	for i := range job.PodGroup.Spec.SubGroupPolicy {
-		nt := job.PodGroup.Spec.SubGroupPolicy[i].NetworkTopology
-		if nt != nil && nt.Mode == scheduling.SoftNetworkTopologyMode {
-			klog.V(3).InfoS("Converting SubGroupPolicy soft topology to hard mode",
-				"job", job.UID, "subGroupPolicy", job.PodGroup.Spec.SubGroupPolicy[i].Name,
-				"maxTier", subJobMaxTier)
-			nt.Mode = scheduling.HardNetworkTopologyMode
-			nt.HighestTierAllowed = &subJobMaxTier
-			nt.HighestTierName = ""
-		}
-	}
-
-	// Convert SubJob-level topology (SubJobInfo has its own deep-copied networkTopology).
-	for _, subJob := range job.SubJobs {
-		subJob.ConvertToHardTopology(subJobMaxTier)
-	}
 }
 
 // getNewAllocatedHyperNode Obtain the newly allocated hyperNode for the job in soft topology mode
